Modificaciones a los iconos y constantes de las paginas home, login, register y dashboard
