+++ src/components/headers/HeaderHome.tsx
@@ -15,7 +15,7 @@ export default function HeaderHome({ query }: Props) {
   query >= 400 ? (res = useMediaQuery(query)) : (res = !useMediaQuery(query))
   return (
     <Box sx={{ flexGrow: 1, width: 1, flexWrap: 'wrap' }}>
-      <AppBar position='static' sx={{ backgroundColor: 'common.white' }}>
+      <AppBar position='static' sx={{ background: 'none', boxShadow: 'none' }}>
         <Toolbar>
           <Logo sx={{ mr: 1 }} />
           <Typography variant='h6' component='div' sx={{ flexGrow: 1 }}>
