+++ src/pages/Dashboard.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { Container } from '@mui/material'
+import { Container, Grid } from '@mui/material'
 import AsideBar from '../components/Drawer/AsideBar'
-import HomeCourse from '../components/HomeLogin/HomeCourse'
-export default function Log() {
+import HomePage from '../components/Content/HomePage'
+export default function DashBoard() {
   return (
     <Container
       maxWidth={false}
@@ -10,8 +10,14 @@ export default function Log() {
         height: '100vh',
         padding: '0 !important',
       }}>
-      <AsideBar />
-      <HomeCourse />
+      <Grid container sx={{ display: 'flex' }}>
+        <Grid item>
+          <AsideBar />
+        </Grid>
+        <Grid item xs sx={{ ml: '128px' }}>
+          <HomePage />
+        </Grid>
+      </Grid>
     </Container>
   )
 }
